User 엔티티 추가 및 기존 id 타입 long으로 변경

+++ src/main/java/com/example/sbb/question/QuestionController.java
@@ -37,7 +37,7 @@ public class QuestionController {
     }
 
     @RequestMapping(value = "/detail/{id}")
-    public String detail(Model model, @PathVariable("id") Integer id, AnswerForm answerForm){
+    public String detail(Model model, @PathVariable("id") long id, AnswerForm answerForm){
         Question question = questionService.getQuestion(id);
         model.addAttribute("question",question);
